Differentiate between admins and owners
At the moment, admins and owners are equivalent
except that owners can:

* approve submissions
* view a list of officers
* update officers' roles
