+++ app/views/application/_menu.html.erb
@@ -12,6 +12,9 @@
     <%= link_to t("drafts.index.link"), drafts_path %>
     <%= link_to t("submissions.index.link"), submissions_path %>
     <%= link_to t("menu.new_person"), new_person_path %>
+  <% end %>
+
+  <% if current_officer.try(:owner?) %>
     <%= link_to t("menu.officers"), officers_path %>
   <% end %>
